テストファイルの構造を変更
