+++ Classifiers/BinaryClassifier/Sources/BinaryClassifier/BinaryTrainingResult.swift
