+++ Classifiers/MultiClassClassifier/Sources/MultiClassClassifier/MultiClassTrainingResult.swift
