+++ Classifiers/OvOClassifier/Sources/OvOClassifier/OvOTrainingResult.swift
