+++ Classifiers/OvOClassifier/Tests/TestResources/mouth_open/OvO_mouth_open_1.jpg
